--- tests/Behat/Mink/Driver/ZombieDriverTest.php
@@ -1,60 +1,0 @@
-<?php
-
-namespace Tests\Behat\Mink\Driver;
-
-use Behat\Mink\Driver\NodeJS\Server\ZombieServer;
-use Behat\Mink\Driver\ZombieDriver;
-
-/**
- * @group zombiedriver
- */
-class ZombieDriverTest extends JavascriptDriverTest
-{
-    protected static function getDriver()
-    {
-        $server = new ZombieServer('127.0.0.1', 8124, 'node');
-
-        return new ZombieDriver($server);
-    }
-
-    public function testWait()
-    {
-        $this->markTestSkipped('Zombie automatically waits for events to fire, so the wait test is irrelevant');
-    }
-
-    public function testIssue193()
-    {
-        $this->markTestSkipped('Zombie.js doesn\'t handle SELECT without values');
-    }
-
-    public function testHtml5FormInputAttribute()
-    {
-        $this->markTestSkipped('Zombie.js doesn\'t HTML5 form attributes. See https://github.com/assaf/zombie/issues/635');
-    }
-
-    public function testHtml5FormButtonAttribute()
-    {
-        $this->markTestSkipped('Zombie.js doesn\'t HTML5 form attributes. See https://github.com/assaf/zombie/issues/635');
-    }
-
-    public function testHtml5FormOutside()
-    {
-        $this->markTestSkipped('Zombie.js doesn\'t HTML5 form attributes. See https://github.com/assaf/zombie/issues/635');
-    }
-
-    public function testSetUserAgent()
-    {
-        $session = $this->getSession();
-
-        $session->setRequestHeader('user-agent', 'foo bar');
-        $session->visit($this->pathTo('/headers.php'));
-        $this->assertContains('foo bar', $session->getPage()->getText());
-    }
-
-    public function testSetRequestHeader()
-    {
-        $this->getSession()->setRequestHeader('foo', 'bar');
-        $this->getSession()->visit($this->pathTo('/headers.php'));
-        $this->assertContains('[HTTP_FOO] => bar', $this->getSession()->getPage()->getText());
-    }
-}
